pyproject.toml
[project]
name = "gandu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.5",
    "groq>=1.1.2",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.1",
    "pdfplumber>=0.11.9",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "tenacity>=9.1.4",
]
